requirements.txt
apscheduler
mysql-connector-python
redis==3.5.3